workflows: add reusable workflows to check PRs for DCO signoffs
If every commit in a PR has not been signed off by the author, fail CI,
and also add a comment explaining what the contributor needs to assent to
and how to do it.  If every commit has been signed off, add a comment
explaining what the contributor has assented to, to ensure that the
implications of a signoff are clear.  If such a comment already exists
when the PR is updated, modify the existing comment rather than adding
a new one.

Do the check in two parts: an unprivileged workflow that checks each
commit and a privileged workflow that adds the PR comment.  The latter
is done separately to avoid any possible privilege escalation from an
untrusted PR.

Individual repositories need to explicitly invoke these reusable
workflows; see the block comments for details.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>