workflows/dco-report: use openslide-bot token to comment
We've been requiring PR write access to the repo, which allows us to
arbitrarily modify PRs, but all we need to do is create or update a
comment.  That can be done by a separate bot account that's completely
unprivileged.

Use the openslide-bot account token if available, and otherwise don't
post a comment.  This also prevents the workflow from commenting on PRs
in forks, which is useful to reduce noise.

Signed-off-by: Benjamin Gilbert <bgilbert@cs.cmu.edu>